# adjust to use recordparser in parallel