eventually this code, as well all methods support, will be integreted in the library.
I know there is a lot of boilerplate here that users will not write themselves in a normal use case, but since this performance can be achieved by user when really needed, I think it is a valid benchmark after all.