@joelverhagen 2021-12-09
> I know there is a lot of boilerplate here that users will not write themselves in a normal use case, but since this performance can be achieved by user when really needed, I think it is a valid benchmark after all.

No worries at all :) I'll let you as the owner of this test to draw the line between code that is written by the user vs. what's in the library.

> if possible, I would like to see a result benchmark before merge.

Latest results (before merge) are available here: https://www.joelverhagen.com/blog/2020/12/fastest-net-csv-parsers

Looks like the unit test is failing. Could you take a look?