@joelverhagen 2021-12-11
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
AMD Ryzen 9 3950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.100
  [Host]     : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
  Job-VNHKCQ : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT

InvocationCount=1  IterationCount=6  LaunchCount=1  
UnrollFactor=1  WarmupCount=2  

```

### BEFORE

|          Method |    Mean |    Error |   StdDev |      Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|---------------- |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
|       CsvHelper | 2.340 s | 0.0300 s | 0.0107 s | 33000.0000 | 18000.0000 | 3000.0000 |    261 MB |
|    RecordParser | **1.781 s** | 0.0304 s | 0.0108 s | 33000.0000 | 18000.0000 | 3000.0000 |    **261 MB** |
| Sylvan_Data_Csv | 1.370 s | 0.0549 s | 0.0196 s | 33000.0000 | 18000.0000 | 3000.0000 |    261 MB |

### AFTER

|          Method |    Mean |    Error |   StdDev |      Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|---------------- |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
|       CsvHelper | 2.323 s | 0.0267 s | 0.0095 s | 33000.0000 | 18000.0000 | 3000.0000 |    261 MB |
|    RecordParser | **1.300 s** | 0.0476 s | 0.0124 s | 37000.0000 | 19000.0000 | 3000.0000 |    **381 MB** |
| Sylvan_Data_Csv | 1.339 s | 0.0377 s | 0.0098 s | 33000.0000 | 18000.0000 | 3000.0000 |    261 MB |