@electricessence 2021-12-11
@joelverhagen @leandromoh Isn't this an unfair comparison?  Am I missing something?
If you parallelize one lib, then you'll need to offer parallel equivalents to all libs.
Some will do better than others at parallel processing.
The implementation functions should be as close to identical as possible to ensure a true apples to apples comparison.
That doesn't stop anyone from adding some queuing or partitioning under the hood but in general isn't this benchmark really asking a simple question? "Which is faster at synchronous processing?"

Thoughts?