@leandromoh 2021-12-12
> @joelverhagen @leandromoh Isn't this an unfair comparison? Am I missing something? If you parallelize one lib, then you'll need to offer parallel equivalents to all libs. Some will do better than others at parallel processing. 

Not necessarily, Usually this kind of libraries process files in a sequential manner. They dont have the feature of be thread-safe, and may not able to parse in parallel. So you can't get the "parallel equivalents to all libs" because it may not exist, but of course someone can open a PR for some library to run in parallel if the feature is an option.

> The implementation functions should be as close to identical as possible to ensure a true apples to apples comparison. That doesn't stop anyone from adding some queuing or partitioning under the hood but in general isn't this benchmark really asking a simple question? "Which is faster at synchronous processing?"
> 
> Thoughts?

In the beginning of the [post](https://www.joelverhagen.com/blog/2020/12/fastest-net-csv-parsers) is said that the benchamark has "_Specific purpose tested: My goal was to find the fastest low-level CSV parser._". Seems me that we don't care how things are done, but how fast things are done.