@leandromoh 2021-12-12
> Let me know if you want to merge as-is.

I did a new commit #7d899d0, I imagine that it result may be a bit faster than previous.