@electricessence 2021-12-12
Looking at the 'implementation' code in the commit I find it overly complex compared to the others.
You might as well change the string.split implementation as well to facilitate paralleling the split operation.
I have a variety of async and even parallel friendly methods but I haven't asked to change the implementation because I think it defeats the purpose of the benchmark.

I challenge this implementation as a quoted string with a newline character is an outlier that easily breaks line segmented processing.  I rewrote my entire lib to handle these cases and when you can't rely on a newline character as a hard segmentation it's nearly impossible to parallelize parsing.  You can however process the string to type conversion in other threads which can definitely improve the real world result.

All that said, if implementing the parser is multiples more lines of code than just using a straight synchronous method, then I don't think that implementation is valid for these benchmarks.  I would ask @leandromoh to provide a separate method that handles the partitioning/etc so that the implementation is as simple as any other.  Using multiple threads to speed things up is totally ok, but if the consumer has to do more work to make it happen, it doesn't seem appropriate here.