@electricessence 2021-12-12
Sorry, I have a hard time not seeing this as straight up cheating.
https://github.com/joelverhagen/NCsvPerf/pull/45/commits/1fa86d43d826bac80b504448ba25efe3148148a5
If you're gonna do this, you should be also doing it for string.split as the baseline.